Modify: search store 수정 - store에 관리하는 상태가 늘어나면서 이름 변경 - 페이지 및 isbn 전역 상태 관리 추가
