--- src/stores/searchResult.ts
@@ -1,14 +1,0 @@
-import { create } from "zustand";
-
-interface State {
-  query: string;
-}
-
-interface Action {
-  setQuery: (query: string) => void;
-}
-
-export const useSearchResult = create<State & Action>((set) => ({
-  query: "",
-  setQuery: (query) => set({ query }),
-}));
